Moved json react component to ui-components
