+++ packages/metadata-common/package.json
@@ -40,11 +40,9 @@
     "@lumino/algorithm": "^1.2.3",
     "@lumino/messaging": "^1.3.3",
     "@lumino/widgets": "^1.11.1",
-    "react": "^16.8.6",
-    "react-json-tree": "^0.11.2"
+    "react": "^16.8.6"
   },
   "devDependencies": {
-    "@types/react-json-tree": "^0.6.11",
     "rimraf": "~3.0.0",
     "typescript": "~3.7.3"
   },
