+++ packages/metadata-common/src/MetadataWidget.tsx
@@ -15,7 +15,11 @@
  */
 
 import { IDictionary, FrontendServices } from '@elyra/application';
-import { ExpandableComponent, trashIcon } from '@elyra/ui-components';
+import {
+  ExpandableComponent,
+  JSONComponent,
+  trashIcon
+} from '@elyra/ui-components';
 
 import { JupyterFrontEnd } from '@jupyterlab/application';
 import {
@@ -28,7 +32,6 @@ import { addIcon, editIcon, LabIcon } from '@jupyterlab/ui-components';
 import { Message } from '@lumino/messaging';
 import { Signal } from '@lumino/signaling';
 import React from 'react';
-import JSONTree from 'react-json-tree';
 
 /**
  * The CSS class added to metadata widgets.
@@ -40,28 +43,6 @@ const METADATA_JSON_CLASS = 'jp-RenderedJSON CodeMirror cm-s-jupyter';
 
 const commands = {
   OPEN_METADATA_EDITOR: 'elyra-metadata-editor:open'
-};
-
-// Provide an invalid theme object (this is on purpose!) to invalidate the
-// react-json-tree's inline styles that override CodeMirror CSS classes
-const theme = {
-  scheme: 'jupyter',
-  base00: 'invalid',
-  base01: 'invalid',
-  base02: 'invalid',
-  base03: 'invalid',
-  base04: 'invalid',
-  base05: 'invalid',
-  base06: 'invalid',
-  base07: 'invalid',
-  base08: 'invalid',
-  base09: 'invalid',
-  base0A: 'invalid',
-  base0B: 'invalid',
-  base0C: 'invalid',
-  base0D: 'invalid',
-  base0E: 'invalid',
-  base0F: 'invalid'
 };
 
 export interface IMetadata {
@@ -139,43 +120,7 @@ export class MetadataDisplay<
   renderExpandableContent(metadata: IDictionary<any>): JSX.Element {
     return (
       <div className={METADATA_JSON_CLASS}>
-        <JSONTree
-          data={metadata.metadata}
-          theme={{
-            extend: theme,
-            valueLabel: 'cm-variable',
-            valueText: 'cm-string',
-            nestedNodeItemString: 'cm-comment'
-          }}
-          invertTheme={false}
-          keyPath={['metadata']}
-          getItemString={(type, data, itemType, itemString) =>
-            Array.isArray(data) ? (
-              // Always display array type and the number of items i.e. "[] 2 items".
-              <span>
-                {itemType} {itemString}
-              </span>
-            ) : Object.keys(data).length === 0 ? (
-              // Only display object type when it's empty i.e. "{}".
-              <span>{itemType}</span>
-            ) : (
-              null! // Upstream typings don't accept null, but it should be ok
-            )
-          }
-          labelRenderer={([label, type]) => {
-            return <span className="cm-keyword">{`${label}: `}</span>;
-          }}
-          valueRenderer={raw => {
-            let className = 'cm-string';
-            if (typeof raw === 'number') {
-              className = 'cm-number';
-            }
-            if (raw === 'true' || raw === 'false') {
-              className = 'cm-keyword';
-            }
-            return <span className={className}>{`${raw}`}</span>;
-          }}
-        />
+        <JSONComponent json={metadata.metadata} root={'metadata'} />
       </div>
     );
   }
