+++ packages/ui-components/package.json
@@ -34,9 +34,11 @@
     "@blueprintjs/select": "^3.11.2",
     "@jupyterlab/ui-components": "^2.0.2",
     "react": "^16.8.6",
-    "react-dom": "^16.8.6"
+    "react-dom": "^16.8.6",
+    "react-json-tree": "^0.11.2"
   },
   "devDependencies": {
+    "@types/react-json-tree": "^0.6.11",
     "rimraf": "~3.0.0",
     "typescript": "~3.7.3"
   },
